Upgrade deps to latest, including react-celo for WC v2 Update metamask firefox fix to newer version Refactor app dir for Next upgrade Remove some unused libs
