+++ jest.config.ts
@@ -4,7 +4,6 @@ const config: JestConfigWithTsJest = {
   preset: 'ts-jest',
   setupFilesAfterEnv: ['<rootDir>/setupTests.ts'],
   transform: {
-    '.+\\.(css|styl|less|sass|scss)$': 'jest-css-modules-transform',
     '.+\\.(ts|tsx)$': ['ts-jest', { tsconfig: 'tsconfig.test.json' }],
   },
   moduleNameMapper: {
