+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@mento-protocol/mento-web",
-  "version": "1.5.1",
+  "version": "1.6.0",
   "description": "A simple DApp for Celo Mento exchanges",
   "keywords": [
     "Celo",
@@ -26,23 +26,21 @@
     "prettier": "prettier --write ./src"
   },
   "dependencies": {
-    "@celo/contractkit": "^3.1.0",
-    "@celo/react-celo": "^4.3.0",
-    "@ethersproject/abi": "^5.5.0",
-    "@ethersproject/address": "^5.5.0",
-    "@ethersproject/units": "^5.5.0",
-    "@headlessui/react": "^1.7.8",
-    "@metamask/inpage-provider": "6.0.1",
+    "@celo/contractkit": "^3.2.0",
+    "@celo/react-celo": "^5.0.3",
+    "@mento-protocol/mento-sdk": "^0.1.1",
     "@metamask/jazzicon": "https://github.com/jmrossy/jazzicon#7a8df28",
-    "@reduxjs/toolkit": "^1.9.2",
+    "@metamask/post-message-stream": "6.1.0",
+    "@metamask/providers": "10.2.1",
+    "@reduxjs/toolkit": "^1.9.3",
     "bignumber.js": "^9.1.1",
+    "ethers": "^5.7.2",
     "formik": "^2.2.9",
     "frappe-charts": "^1.6.2",
-    "next": "13.1.6",
+    "next": "13.2.1",
     "next-persist": "^1.2.4",
-    "post-message-stream": "3.0.0",
     "react": "^18.2.0",
-    "react-accessible-dropdown-menu-hook": "^3.2.0",
+    "react-accessible-dropdown-menu-hook": "^4.0.0",
     "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
     "react-redux": "^8.0.5",
@@ -50,31 +48,27 @@
     "react-toastify": "^9.1.1"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@trivago/prettier-plugin-sort-imports": "^4.0.0",
+    "@trivago/prettier-plugin-sort-imports": "^4.1.1",
     "@types/jest": "^29.4.0",
-    "@types/node": "^18.11.18",
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
+    "@types/node": "^18.14.1",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "@types/react-modal": "^3.13.1",
     "@types/react-redux": "^7.1.25",
     "@types/react-select": "4.0.18",
-    "@typescript-eslint/eslint-plugin": "^5.49.0",
-    "@typescript-eslint/parser": "^5.49.0",
+    "@typescript-eslint/eslint-plugin": "^5.53.0",
+    "@typescript-eslint/parser": "^5.53.0",
     "autoprefixer": "^10.4.13",
-    "eslint": "^8.33.0",
-    "eslint-config-next": "^13.1.6",
+    "eslint": "^8.34.0",
+    "eslint-config-next": "^13.2.1",
     "eslint-config-prettier": "^8.6.0",
-    "jest": "^29.4.1",
-    "jest-css-modules-transform": "^4.4.2",
+    "jest": "^29.4.3",
     "postcss": "^8.4.21",
-    "prettier": "^2.8.3",
-    "tailwindcss": "^3.2.4",
+    "prettier": "^2.8.4",
+    "tailwindcss": "^3.2.7",
     "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.4"
+    "typescript": "^4.9.5"
   },
   "resolutions": {
     "bignumber.js": "9.1.1",
