+++ src/components/nav/Footer.tsx
@@ -1,10 +1,10 @@
 import BigNumber from 'bignumber.js'
 import Image from 'next/image'
 import { useState } from 'react'
-import { useAppSelector } from 'src/app/hooks'
 import { NetworkModal } from 'src/components/nav/NetworkModal'
 import { config } from 'src/config/config'
 import { STALE_BLOCK_TIME } from 'src/config/consts'
+import { useAppSelector } from 'src/features/store/hooks'
 import Discord from 'src/images/logos/discord.svg'
 import Github from 'src/images/logos/github.svg'
 import Twitter from 'src/images/logos/twitter.svg'
