+++ src/components/nav/NetworkModal.tsx
@@ -1,12 +1,12 @@
 import { Alfajores, Baklava, Mainnet, Network, useCelo } from '@celo/react-celo'
 import ReactModal from 'react-modal'
 import { toast } from 'react-toastify'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { IconButton } from 'src/components/buttons/IconButton'
 import { reset as accountReset } from 'src/features/accounts/accountSlice'
 import { reset as blockReset } from 'src/features/blocks/blockSlice'
 import { resetTokenPrices } from 'src/features/chart/tokenPriceSlice'
 import { reset as grandaReset } from 'src/features/granda/grandaSlice'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { reset as swapReset } from 'src/features/swap/swapSlice'
 import XCircle from 'src/images/icons/x-circle.svg'
 import { HrDivider } from 'src/layout/HrDivider'
