+++ src/config/config.ts
@@ -5,10 +5,12 @@ interface Config {
   discordUrl: string
   blockscoutUrl: string
   showPriceChart: boolean
+  walletConnectProjectId?: string
 }
 
 const isDevMode = process?.env?.NODE_ENV === 'development'
 const version = process?.env?.NEXT_PUBLIC_VERSION ?? null
+const walletConnectProjectId = process?.env?.WALLET_CONNECT_ID
 
 export const config: Config = Object.freeze({
   debug: isDevMode,
@@ -17,4 +19,5 @@ export const config: Config = Object.freeze({
   discordUrl: 'https://discord.com/invite/Zszgng9NdF',
   blockscoutUrl: 'https://explorer.celo.org',
   showPriceChart: false,
+  walletConnectProjectId,
 })
