+++ src/features/chart/PriceChartCelo.tsx
@@ -1,11 +1,11 @@
 import { Mainnet, useCelo } from '@celo/react-celo'
 import { useEffect } from 'react'
 import { toast } from 'react-toastify'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { NativeTokenId } from 'src/config/tokens'
 import styles from 'src/features/chart/PriceChart.module.css'
 import { fetchTokenPrice } from 'src/features/chart/fetchPrices'
 import { tokenPriceHistoryToChartData } from 'src/features/chart/utils'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { FloatingBox } from 'src/layout/FloatingBox'
 import { Color } from 'src/styles/Color'
 import { logger } from 'src/utils/logger'
