+++ src/features/granda/ProposalConfirm.tsx
@@ -4,7 +4,6 @@ import { useCelo } from '@celo/react-celo'
 import Image from 'next/image'
 import { useEffect } from 'react'
 import { toast } from 'react-toastify'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { toastToYourSuccess } from 'src/components/TxSuccessToast'
 import { BackButton } from 'src/components/buttons/BackButton'
 import { RefreshButton } from 'src/components/buttons/RefreshButton'
@@ -18,6 +17,7 @@ import { fetchOracleRates } from 'src/features/granda/fetchOracleRates'
 import { fetchProposals } from 'src/features/granda/fetchProposals'
 import { setFormValues } from 'src/features/granda/grandaSlice'
 import { getExchangeValues } from 'src/features/granda/utils'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { SwapConfirmSummary } from 'src/features/swap/SwapConfirm'
 import InfoCircle from 'src/images/icons/info-circle.svg'
 import { FloatingBox } from 'src/layout/FloatingBox'
