+++ src/features/granda/ProposalForm.tsx
@@ -1,9 +1,9 @@
 import Image from 'next/image'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { TextLink } from 'src/components/buttons/TextLink'
 import { setFormValues, setSubpage } from 'src/features/granda/grandaSlice'
 import { GrandaSubpage } from 'src/features/granda/types'
 import { getExchangeValues } from 'src/features/granda/utils'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { SwapFormInner } from 'src/features/swap/SwapForm'
 import { SwapFormValues } from 'src/features/swap/types'
 import { useFormValidator } from 'src/features/swap/useFormValidator'
