+++ src/features/granda/ProposalList.tsx
@@ -1,6 +1,5 @@
 import BigNumber from 'bignumber.js'
 import Image from 'next/image'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { Spinner } from 'src/components/animation/Spinner'
 import { BackButton } from 'src/components/buttons/BackButton'
 import { IconButton } from 'src/components/buttons/IconButton'
@@ -8,6 +7,7 @@ import { SolidButton } from 'src/components/buttons/SolidButton'
 import { CELO, NativeTokens } from 'src/config/tokens'
 import { setSubpage, viewProposal } from 'src/features/granda/grandaSlice'
 import { GrandaProposal, GrandaProposalState, GrandaSubpage } from 'src/features/granda/types'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import ArrowRight from 'src/images/icons/arrow-right-short.svg'
 import PlusCircle from 'src/images/icons/plus-circle-fill.svg'
 import { TokenIcon } from 'src/images/tokens/TokenIcon'
