+++ src/features/granda/ProposalView.tsx
@@ -4,7 +4,6 @@ import BigNumber from 'bignumber.js'
 import { useEffect, useState } from 'react'
 import ReactModal from 'react-modal'
 import { toast } from 'react-toastify'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { toastToYourSuccess } from 'src/components/TxSuccessToast'
 import { BackButton } from 'src/components/buttons/BackButton'
 import { IconButton } from 'src/components/buttons/IconButton'
@@ -17,6 +16,7 @@ import { getGrandaMento } from 'src/contract-wrappers/granda-mento'
 import { fetchProposals } from 'src/features/granda/fetchProposals'
 import { clearProposal } from 'src/features/granda/grandaSlice'
 import { GrandaProposal, GrandaProposalState } from 'src/features/granda/types'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { SwapConfirmSummary } from 'src/features/swap/SwapConfirm'
 import XCircle from 'src/images/icons/x-circle.svg'
 import { FloatingBox } from 'src/layout/FloatingBox'
