+++ src/features/granda/fetchConfig.ts
@@ -1,9 +1,9 @@
 import type { MiniContractKit } from '@celo/contractkit/lib/mini-kit'
 import { createAsyncThunk } from '@reduxjs/toolkit'
-import type { AppDispatch, AppState } from 'src/app/store'
 import { kitContractToNativeToken } from 'src/config/tokenMapping'
 import { getGrandaMento } from 'src/contract-wrappers/granda-mento'
 import { GrandaConfig, SizeLimits } from 'src/features/granda/types'
+import type { AppDispatch, AppState } from 'src/features/store/store'
 import { isValidAddress } from 'src/utils/addresses'
 import { toWei } from 'src/utils/amount'
 import { logger } from 'src/utils/logger'
