+++ src/features/granda/fetchOracleRates.ts
@@ -1,11 +1,11 @@
 import type { MiniContractKit } from '@celo/contractkit/lib/mini-kit'
 import { createAsyncThunk } from '@reduxjs/toolkit'
-import type { AppDispatch, AppState } from 'src/app/store'
 import { MAX_EXCHANGE_RATE, MIN_EXCHANGE_RATE } from 'src/config/consts'
 import { nativeTokenToKitToken } from 'src/config/tokenMapping'
 import { NativeTokenId, StableTokenIds } from 'src/config/tokens'
 import { getSortedOracles } from 'src/contract-wrappers/sorted-oracles'
 import { OracleRates } from 'src/features/granda/types'
+import type { AppDispatch, AppState } from 'src/features/store/store'
 import { SimpleExchangeRate } from 'src/features/swap/types'
 import { logger } from 'src/utils/logger'
 import { areRatesStale } from 'src/utils/time'
