+++ src/features/granda/fetchProposals.ts
@@ -2,12 +2,12 @@ import type { CeloTokenType } from '@celo/contractkit'
 import type { MiniContractKit } from '@celo/contractkit/lib/mini-kit'
 import { ExchangeProposalState } from '@celo/contractkit/lib/wrappers/GrandaMento'
 import { createAsyncThunk } from '@reduxjs/toolkit'
-import type { AppDispatch, AppState } from 'src/app/store'
 import { GRANDA_PROPOSAL_STALE_TIME } from 'src/config/consts'
 import { kitTokenToNativeToken } from 'src/config/tokenMapping'
 import { NativeTokenId, isStableToken } from 'src/config/tokens'
 import { getGrandaMento } from 'src/contract-wrappers/granda-mento'
 import { GrandaProposal, GrandaProposalState } from 'src/features/granda/types'
+import type { AppDispatch, AppState } from 'src/features/store/store'
 import { areAddressesEqual, isValidAddress } from 'src/utils/addresses'
 import { logger } from 'src/utils/logger'
 import { isStale } from 'src/utils/time'
