+++ src/features/polling/PollingWorker.tsx
@@ -1,13 +1,13 @@
 import { useCelo } from '@celo/react-celo'
 import { useEffect } from 'react'
 import { toast } from 'react-toastify'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { STATUS_POLLER_DELAY } from 'src/config/consts'
 import { fetchBalances } from 'src/features/accounts/fetchBalances'
 import { fetchLatestBlock } from 'src/features/blocks/fetchLatestBlock'
 import { fetchConfig } from 'src/features/granda/fetchConfig'
 import { fetchOracleRates } from 'src/features/granda/fetchOracleRates'
 import { fetchProposals } from 'src/features/granda/fetchProposals'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { fetchExchangeRates } from 'src/features/swap/fetchExchangeRates'
 import { logger } from 'src/utils/logger'
 import { useInterval } from 'src/utils/timeout'
