+++ src/features/swap/SwapConfirm.tsx
@@ -3,7 +3,6 @@ import { useCelo } from '@celo/react-celo'
 import BigNumber from 'bignumber.js'
 import { useEffect } from 'react'
 import { toast } from 'react-toastify'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { toastToYourSuccess } from 'src/components/TxSuccessToast'
 import { BackButton } from 'src/components/buttons/BackButton'
 import { RefreshButton } from 'src/components/buttons/RefreshButton'
@@ -17,6 +16,7 @@ import {
 import { getExchangeContract, getTokenContract } from 'src/config/tokenMapping'
 import { NativeTokenId, NativeTokens } from 'src/config/tokens'
 import { fetchBalances } from 'src/features/accounts/fetchBalances'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { fetchExchangeRates } from 'src/features/swap/fetchExchangeRates'
 import { setFormValues } from 'src/features/swap/swapSlice'
 import { SwapFormValues } from 'src/features/swap/types'
