+++ src/features/swap/SwapForm.tsx
@@ -2,13 +2,13 @@ import { Connector, useCelo } from '@celo/react-celo'
 import { Field, Form, Formik, FormikErrors, useFormikContext } from 'formik'
 import { useCallback } from 'react'
 import { toast } from 'react-toastify'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { IconButton } from 'src/components/buttons/IconButton'
 import { SolidButton } from 'src/components/buttons/SolidButton'
 import { RadioInput } from 'src/components/input/RadioInput'
 import TokenSelectField, { TokenOption } from 'src/components/input/TokenSelectField'
 import { CELO, NativeTokenId, cEUR, cREAL, cUSD, isStableToken } from 'src/config/tokens'
 import { AccountBalances } from 'src/features/accounts/fetchBalances'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 import { SettingsMenu } from 'src/features/swap/SettingsMenu'
 import { setFormValues } from 'src/features/swap/swapSlice'
 import { SwapFormValues } from 'src/features/swap/types'
