+++ src/features/swap/fetchExchangeRates.ts
@@ -1,9 +1,9 @@
 import type { MiniContractKit } from '@celo/contractkit/lib/mini-kit'
 import { createAsyncThunk } from '@reduxjs/toolkit'
-import type { AppDispatch, AppState } from 'src/app/store'
 import { MAX_EXCHANGE_SPREAD } from 'src/config/consts'
 import { getExchangeContract } from 'src/config/tokenMapping'
 import { NativeTokenId, StableTokenIds } from 'src/config/tokens'
+import type { AppDispatch, AppState } from 'src/features/store/store'
 import { ExchangeRate, ToCeloRates } from 'src/features/swap/types'
 import { logger } from 'src/utils/logger'
 import { areRatesStale } from 'src/utils/time'
