+++ src/pages/500.tsx
@@ -1,4 +1,4 @@
-import { FailContent } from 'src/app/FailScreen'
+import { FailContent } from 'src/components/Errors'
 
 export default function Custom500() {
   return <FailContent details="500 - Server-side error occurred" />
