+++ src/pages/granda.tsx
@@ -1,11 +1,11 @@
 import { useEffect } from 'react'
-import { useAppDispatch, useAppSelector } from 'src/app/hooks'
 import { ProposalConfirm } from 'src/features/granda/ProposalConfirm'
 import { ProposalForm } from 'src/features/granda/ProposalForm'
 import { ProposalList } from 'src/features/granda/ProposalList'
 import { ProposalView } from 'src/features/granda/ProposalView'
 import { activateGranda, setSubpage } from 'src/features/granda/grandaSlice'
 import { GrandaSubpage } from 'src/features/granda/types'
+import { useAppDispatch, useAppSelector } from 'src/features/store/hooks'
 
 export default function GrandaPage() {
   const subpage = useAppSelector((s) => s.granda.subpage)
